test(matrix): isolate account migration proof
